requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.9.1
    # via django
django==5.2.6
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
sqlparse==0.5.3
    # via django
